[ #31 ] More functions

+++ core/src/org/ice1000/jimgui/JImFont.java
@@ -25,7 +25,7 @@ public final class JImFont extends JImGuiFontGen {
 	}
 
 	public void setFallbackChar(char wChar) {
-		setFallbackChar((short) wChar);
+		setFallbackChar((int) wChar);
 	}
 
 	public @NotNull JImFontAtlas getContainerAtlas() {
